Implement cross mapping on CPU

+++ src/dataset.h
@@ -8,20 +8,20 @@ class Timeseries
 {
 public:
     Timeseries() : Timeseries(nullptr, 0) {}
-    Timeseries(const float *data, size_t size) : _data(data), _size(size) {}
+    Timeseries(const float *data, uint32_t size) : _data(data), _size(size) {}
     Timeseries(const std::vector<float> &vec)
         : Timeseries(vec.data(), vec.size())
     {
     }
 
     const float *data() const { return _data; }
-    size_t size() const { return _size; }
+    uint32_t size() const { return _size; }
 
-    const float &operator[](size_t i) const { return _data[i]; };
+    const float &operator[](uint32_t i) const { return _data[i]; };
 
 protected:
     const float *_data;
-    size_t _size;
+    uint32_t _size;
 };
 
 class Dataset
@@ -32,11 +32,11 @@ public:
     Dataset() : _n_rows(0), is_header(true) {}
     Dataset(const std::string &path);
 
-    size_t n_rows() const { return _n_rows; }
+    uint32_t n_rows() const { return _n_rows; }
 
 protected:
     std::vector<std::vector<float>> columns;
-    size_t _n_rows;
+    uint32_t _n_rows;
     bool is_header;
 };
